test(load): add test for extends taking precedence over local
This bug is mainly to document the bug, not for fixing it. The best way I could
think of to document this bug is to add a test for it.

The bug is that after we merge in the local config, we will re-run the extend
function. This is suitable if the local config also has an extend key and
we'd want to merge those config files as well.

However, I think we should not re-run the extend function on the extended files
already merged in from the global config. Yet, we do. As such, the extended
files from the global config will take precedence over the local config.